requirements.txt
langchain
flask
python-dotenv
langchain-pinecone
langchain_community
langchain_openai
langchain_experimental
from-root
pandas
-e .